Refactor getAddress and move getAddressFromPublicKey to convert
