+++ test/crypto/keys.js
@@ -15,7 +15,6 @@
 import {
 	getPrivateAndPublicKeyFromSecret,
 	getRawPrivateAndPublicKeyFromSecret,
-	getAddressFromPublicKey,
 	getKeys,
 } from '../../src/crypto/keys';
 import {
@@ -24,7 +23,6 @@ import {
 
 describe('keys', () => {
 	const defaultSecret = 'secret';
-	const defaultAddress = '18160565574430594874L';
 	const expectedPublicKey = '5d036a858ce89f844491762eb89e2bfbd50a4a0a0da658e4b2628b25b117ae09';
 	const expectedPrivateKey = '2bb80d537b1da3e38bd30361aa855686bde0eacd7162fef6a25fe97bf527a25b5d036a858ce89f844491762eb89e2bfbd50a4a0a0da658e4b2628b25b117ae09';
 	describe('#getPrivateAndPublicKeyFromSecret', () => {
@@ -51,14 +49,6 @@ describe('keys', () => {
 		it('should create buffer privateKey', () => {
 			(bufferToHex(Buffer.from(keypair.privateKey)))
 				.should.be.equal(expectedPrivateKey);
-		});
-	});
-
-	describe('#getAddressFromPublicKey', () => {
-		const address = getAddressFromPublicKey(expectedPublicKey);
-
-		it('should generate address from publicKey', () => {
-			(address).should.be.equal(defaultAddress);
 		});
 	});
